Run lasso estimation on the prepared matrixes

Run lasso estimation on the prepared matrixes

estimate_matches(source_x_mat, source_y_mat, source_w_mat, target_x_mat,
  n_near = 5, n_folds = 10, force_lambda = NULL, reduced = TRUE)

Arguments

source_x_mat, source_y_mat, source_w_mat
matrixes of independent, dependent variables and weights

target_x_mat
matrixes of independent variables for the prediction sample

n_near
number of the nearest observations to derive a random match. If `n_near` is greater than `length(match_vector)`, minimum out of two is used to create a sample for selecting a random match value.

n_folds
number of folds for cross-validation

force_lambda
allows to specify one lamda value. Shoul be 0, when we want to switch to the linear regression.

reduced
if TRUE terurns reduced outpur without specific regression details.

Value

In both reduced=TRUE and reduced=FALSE forms, the function returns a list with elements. In the form reduced=TRUE only results of matching are returned:

source_y_hat is the vecrtor of predicted values based on the result of the lasso regression with nfolds cross validation and "mse" measure for identifying minimizing value of lambda. It uses source_x_mat, source_y_mat, and source_w_mat for running regression and predictind source_y_hat.

target_y_hat is the vetor of predicted values produced using the source_y_hat regression results and target_x_mat independent variables matrix.

match is the dataframe with: columns target_id - index of each target_y_hat value; column target_y_hat - its' value; column source_id - position of the nearest match from the source_y_hat vector and column source_y_hat - value of the nearest match

In the not reduced form the list contains more elements:

items source_x_mat, source_y_mat, source_w_mat and target_x_mat from the inputs to the functions.

lambda_cv - result of the cv.glmnet.

fit - result of the glmnet

Examples

library(dplyr)

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

library(purrr)
library(glmnet)

#> Loading required package: Matrix

#> Loading required package: foreach

#> 
#> Attaching package: 'foreach'

#> The following objects are masked from 'package:purrr':
#> 
#>     accumulate, when

#> Loaded glmnet 2.0-16

library(lassopmm)

# Run estimate of a fake data
XX <- as.matrix(mtcars[, !names(mtcars) %in% "hp"])
YY <- as.matrix(mtcars[, "hp"])
WW <- matrix(rep(1, nrow(mtcars)), ncol = 1, nrow = nrow(mtcars))
XX1 <- as.matrix(mtcars[1:10, !names(mtcars) %in% "hp"])

# Running simple estimation and returning
a <- estimate_matches(source_x_mat = XX, source_y_mat = YY, source_w_mat = WW, target_x_mat = XX1, reduced = FALSE, n_near = 5)

# Extract regression coefficients
a$fit %>% coef()

#> 11 x 1 sparse Matrix of class "dgCMatrix"
#>                      s0
#> (Intercept) 148.1928176
#> mpg          -0.1106935
#> cyl           3.5699209
#> disp          0.2339600
#> drat          .        
#> wt            .        
#> qsec         -6.5904714
#> vs            .        
#> am            .        
#> gear          .        
#> carb         15.0334872

str(a, max.level = 1)

#> List of 9
#>  $ source_x_mat: num [1:32, 1:10] 21 21 22.8 21.4 18.7 18.1 14.3 24.4 22.8 19.2 ...
#>   ..- attr(*, "dimnames")=List of 2
#>  $ source_w_mat: num [1:32, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ source_y_mat: num [1:32, 1] 110 110 93 110 175 105 245 62 95 123 ...
#>  $ target_x_mat: num [1:10, 1:10] 21 21 22.8 21.4 18.7 18.1 14.3 24.4 22.8 19.2 ...
#>   ..- attr(*, "dimnames")=List of 2
#>  $ lambda_cv   :List of 10
#>   ..- attr(*, "class")= chr "cv.glmnet"
#>  $ fit         :List of 12
#>   ..- attr(*, "class")= chr [1:2] "elnet" "glmnet"
#>  $ source_y_hat: Named num [1:32] 156.4 152.7 77.6 114.5 176.8 ...
#>   ..- attr(*, "names")= chr [1:32] "Mazda RX4" "Mazda RX4 Wag" "Datsun 710" "Hornet 4 Drive" ...
#>  $ target_y_hat: Named num [1:10] 156.4 152.7 77.6 114.5 176.8 ...
#>   ..- attr(*, "names")= chr [1:10] "Mazda RX4" "Mazda RX4 Wag" "Datsun 710" "Hornet 4 Drive" ...
#>  $ match       :Classes 'tbl_df', 'tbl' and 'data.frame':	10 obs. of  4 variables:

# Developing a sample bootsrtap vector
perm_example <- purrr::map(1:5, ~ sample(1:nrow(YY), nrow(YY), TRUE))

# Running estimations on every single bootstrap permutation vector
a_boot <-
  perm_example %>%
  purrr::map(~ lassopmm::estimate_matches(
    source_x_mat = XX[.x, ],
    source_y_mat = YY[.x, ],
    source_w_mat = WW[.x, ],
    target_x_mat = XX1,
    reduced = FALSE,
    n_near = 5
  ))

# Exploring the structure
a_boot %>%
  str(max.level = 1)

#> List of 5
#>  $ :List of 9
#>  $ :List of 9
#>  $ :List of 9
#>  $ :List of 9
#>  $ :List of 9

# Accesssing fit of a specific bootstrap iteration
a_boot[[3]]$fit

#> 
#> Call:  glmnet::glmnet(x = source_x_mat, y = source_y_mat, weights = source_w_mat,      alpha = 1, lambda = lambda_cv$lambda.min) 
#> 
#>      Df   %Dev Lambda
#> [1,]  4 0.8792  3.662

# Doing the same with each single bootstrap iteration
a_boot %>%
  map("fit")

#> [[1]]
#> 
#> Call:  glmnet::glmnet(x = source_x_mat, y = source_y_mat, weights = source_w_mat,      alpha = 1, lambda = lambda_cv$lambda.min) 
#> 
#>      Df   %Dev Lambda
#> [1,]  4 0.9052  4.007
#> 
#> [[2]]
#> 
#> Call:  glmnet::glmnet(x = source_x_mat, y = source_y_mat, weights = source_w_mat,      alpha = 1, lambda = lambda_cv$lambda.min) 
#> 
#>      Df  %Dev Lambda
#> [1,]  6 0.893  1.775
#> 
#> [[3]]
#> 
#> Call:  glmnet::glmnet(x = source_x_mat, y = source_y_mat, weights = source_w_mat,      alpha = 1, lambda = lambda_cv$lambda.min) 
#> 
#>      Df   %Dev Lambda
#> [1,]  4 0.8792  3.662
#> 
#> [[4]]
#> 
#> Call:  glmnet::glmnet(x = source_x_mat, y = source_y_mat, weights = source_w_mat,      alpha = 1, lambda = lambda_cv$lambda.min) 
#> 
#>      Df   %Dev Lambda
#> [1,]  5 0.8921  3.755
#> 
#> [[5]]
#> 
#> Call:  glmnet::glmnet(x = source_x_mat, y = source_y_mat, weights = source_w_mat,      alpha = 1, lambda = lambda_cv$lambda.min) 
#> 
#>      Df   %Dev Lambda
#> [1,]  7 0.9581 0.9094
#> 

# Extracting coefficinets from each single bootstrap iteration
a_boot %>%
  map("fit") %>%
  map(coefficients)

#> [[1]]
#> 11 x 1 sparse Matrix of class "dgCMatrix"
#>                      s0
#> (Intercept) 247.9566066
#> mpg           .        
#> cyl           .        
#> disp          0.1118758
#> drat        -25.4415333
#> wt            .        
#> qsec         -5.3082590
#> vs            .        
#> am            .        
#> gear          .        
#> carb         21.3545279
#> 
#> [[2]]
#> 11 x 1 sparse Matrix of class "dgCMatrix"
#>                      s0
#> (Intercept) 146.3433278
#> mpg           .        
#> cyl           2.2251952
#> disp          0.2600576
#> drat        -10.5620695
#> wt            .        
#> qsec         -3.8259092
#> vs            .        
#> am           -2.1783548
#> gear          .        
#> carb         10.8622102
#> 
#> [[3]]
#> 11 x 1 sparse Matrix of class "dgCMatrix"
#>                      s0
#> (Intercept) 172.8584444
#> mpg          -0.5950853
#> cyl           .        
#> disp          0.2755696
#> drat          .        
#> wt            .        
#> qsec         -6.5775400
#> vs            .        
#> am            .        
#> gear          .        
#> carb         12.7346563
#> 
#> [[4]]
#> 11 x 1 sparse Matrix of class "dgCMatrix"
#>                      s0
#> (Intercept) 324.0275084
#> mpg          -3.0619630
#> cyl           2.0224922
#> disp          0.1717776
#> drat          .        
#> wt            .        
#> qsec        -10.5715908
#> vs            .        
#> am            .        
#> gear          .        
#> carb          5.2818973
#> 
#> [[5]]
#> 11 x 1 sparse Matrix of class "dgCMatrix"
#>                      s0
#> (Intercept) 145.3589940
#> mpg          -0.2860217
#> cyl          15.9466358
#> disp          0.1671716
#> drat          .        
#> wt            .        
#> qsec        -11.4087759
#> vs           30.5658434
#> am           22.7164708
#> gear          .        
#> carb         18.5586363
#> 

# Combine all coefficients in a table
a_boot %>%
  map("fit") %>%
  map(broom::tidy) %>%
  map(~ select(.x, term, estimate)) %>%
  map2(.y = 1:length(.), ~ rename_at(.x, vars(estimate), list(~ paste0(., "_", .y)))) %>%
  reduce(full_join)

#> Joining, by = "term"

#> Joining, by = "term"

#> Joining, by = "term"

#> Joining, by = "term"

#> # A tibble: 9 x 6
#>   term        estimate_1 estimate_2 estimate_3 estimate_4 estimate_5
#>   <chr>            <dbl>      <dbl>      <dbl>      <dbl>      <dbl>
#> 1 (Intercept)    248.       146.       173.       324.       145.   
#> 2 disp             0.112      0.260      0.276      0.172      0.167
#> 3 drat           -25.4      -10.6       NA         NA         NA    
#> 4 qsec            -5.31      -3.83      -6.58     -10.6      -11.4  
#> 5 carb            21.4       10.9       12.7        5.28      18.6  
#> 6 cyl             NA          2.23      NA          2.02      15.9  
#> 7 am              NA         -2.18      NA         NA         22.7  
#> 8 mpg             NA         NA         -0.595     -3.06      -0.286
#> 9 vs              NA         NA         NA         NA         30.6