Find neares match between a vector of observations and a vector of alternatives

Find neares match between a vector of observations and a vector of alternatives

find_near(observ_vector, match_vector, n_near = 5)

Arguments

observ_vector
named vector or value to find a match from `match_vector`. If the vector has no names function will still work, but the resulting `target_id` variable will be empty.

match_vector
named vector which is used to find a nearest match to the `observ_vector`.

n_near
number of the nearest observations to derive a random match. If `n_near` is greater than `length(match_vector)`, minimum out of two is used to create a sample for selecting a random match value.

Value

A dataframe with the number of rows equall to the lenght of the `observ_vector` vector and one mathcing value and its index from the `match_vector` corresponding to each value afrom the `observ_vector`.

Examples

library(dplyr)
library(purrr)
library(glmnet)
library(lassopmm)

obs <- setNames(rnorm(5), as.character(1:5))
match_v <- setNames(rnorm(50), as.character(1:50))
find_near(obs, match_v, n_near = 10)

#> # A tibble: 5 x 4
#>   target_id target_y_hat source_id source_y_hat
#>   <chr>            <dbl> <chr>            <dbl>
#> 1 1               -1.10  42              -1.19 
#> 2 2               -0.633 43              -0.752
#> 3 3               -2.06  34              -1.43 
#> 4 4                2.65  35               1.38 
#> 5 5               -1.15  30              -0.903