Generates a list of permutation vectors for bootsrapping (old version not used)

Generates a list of permutation vectors for bootsrapping (old version not used)

get_bootstrap_permutation(.data, n_boot = NULL)

Arguments

.data
dataframe containing two columns group and id.

n_boot
number of bootstrap permulations (resamplings). If n_rep = 0, no bootstraping is performed and we have the bootstrap sample, whcih consist of the same observations in the same order as source data.

Examples

library(tibble)
library(dplyr)
library(purrr)
library(lassopmm)

# Generating sample data
n <- 100
sa_df <-
  tibble(
    ids = 1:n,
    .strata = statar::xtile(ids, 10),
    .cluster = c(
      rep(5, n * 0.1),
      rep(2, n * 0.3),
      rep(4, n * 0.4),
      rep(3, n * 0.05),
      rep(1, n * 0.15)
    )
  )

# Dataset, where cluster equall to strata
sa_df1 <- sa_df %>% mutate(.cluster = .strata)

# Dataset, where cluster equall to one number
sa_df2 <- sa_df %>% mutate(.cluster = 0)

# Dataset, where strata equall to one number
sa_df3 <- sa_df %>% mutate(.strata = 1)

# Function for generating cross-tabulation statistics
my_table <- function(x, y) {
  a <- table(x, y)
  b <- cbind(a, t(t(margin.table(a, 1))))
  cc <-
    margin.table(a, 2) %>%
    as.matrix() %>%
    t() %>%
    cbind(., sum(.))
  rbind(b, cc)
}

# Example of simple bootstrapping with clusters and stratas
# Original data strcructure
with(sa_df, my_table(.cluster, .strata))

#>    1  2  3 4  5  6  7  8  9 10    
#> 1  0  0  0 0  0  0  0  0  5 10  15
#> 2  0 10 11 9  0  0  0  0  0  0  30
#> 3  0  0  0 0  0  0  0  0  5  0   5
#> 4  0  0  0 0 10 10 10 10  0  0  40
#> 5 10  0  0 0  0  0  0  0  0  0  10
#>   10 10 11 9 10 10 10 10 10 10 100

# Structure of the three repetedly sampled datasets
get_bootstrap_permutation(sa_df, 3) %>%
  map(~ tibble(ids = .x) %>% left_join(sa_df, "ids")) %>%
  map(~ with(.x, my_table(.cluster, .strata)))

#> [[1]]
#>    1  5  6  7  8 9    
#> 3  0  0  0  0  0 5   5
#> 4  0 20 20 20 20 0  80
#> 5 20  0  0  0  0 0  20
#>   20 20 20 20 20 5 105
#> 
#> [[2]]
#>    2  3 4  5  6  7  8  9    
#> 2 10 11 9  0  0  0  0  0  30
#> 3  0  0 0  0  0  0  0 10  10
#> 4  0  0 0 20 20 20 20  0  80
#>   10 11 9 20 20 20 20 10 120
#> 
#> [[3]]
#>    1  2  3 4  5  6  7  8    
#> 2  0 10 11 9  0  0  0  0  30
#> 4  0  0  0 0 20 20 20 20  80
#> 5 20  0  0 0  0  0  0  0  20
#>   20 10 11 9 20 20 20 20 130
#> 

# Example of sampling with clusters and stratas,
# where data contains clusters equall to stratas.
# This is an equivalent of the only clustered sampling
with(sa_df1, my_table(.cluster, .strata))

#>     1  2  3 4  5  6  7  8  9 10    
#> 1  10  0  0 0  0  0  0  0  0  0  10
#> 2   0 10  0 0  0  0  0  0  0  0  10
#> 3   0  0 11 0  0  0  0  0  0  0  11
#> 4   0  0  0 9  0  0  0  0  0  0   9
#> 5   0  0  0 0 10  0  0  0  0  0  10
#> 6   0  0  0 0  0 10  0  0  0  0  10
#> 7   0  0  0 0  0  0 10  0  0  0  10
#> 8   0  0  0 0  0  0  0 10  0  0  10
#> 9   0  0  0 0  0  0  0  0 10  0  10
#> 10  0  0  0 0  0  0  0  0  0 10  10
#>    10 10 11 9 10 10 10 10 10 10 100

# Structure of the three repetedly sampled datasets
get_bootstrap_permutation(sa_df1, 3) %>%
  map(~ tibble(ids = .x) %>% left_join(sa_df1, "ids")) %>%
  map(~ with(.x, my_table(.cluster, .strata)))

#> [[1]]
#>    1  2  3 4  5  7  8  9    
#> 1 10  0  0 0  0  0  0  0  10
#> 2  0 10  0 0  0  0  0  0  10
#> 3  0  0 11 0  0  0  0  0  11
#> 4  0  0  0 9  0  0  0  0   9
#> 5  0  0  0 0 10  0  0  0  10
#> 7  0  0  0 0  0 10  0  0  10
#> 8  0  0  0 0  0  0 30  0  30
#> 9  0  0  0 0  0  0  0 10  10
#>   10 10 11 9 10 10 30 10 100
#> 
#> [[2]]
#>    2 4  7  8  9   
#> 2 10 0  0  0  0 10
#> 4  0 9  0  0  0  9
#> 7  0 0 30  0  0 30
#> 8  0 0  0 20  0 20
#> 9  0 0  0  0 30 30
#>   10 9 30 20 30 99
#> 
#> [[3]]
#>     3 4  5  8  9 10    
#> 3  11 0  0  0  0  0  11
#> 4   0 9  0  0  0  0   9
#> 5   0 0 20  0  0  0  20
#> 8   0 0  0 20  0  0  20
#> 9   0 0  0  0 10  0  10
#> 10  0 0  0  0  0 30  30
#>    11 9 20 20 10 30 100
#> 

# Example of sampling with clusters and stratas,
# where data contains only one cluster equall to 0
# This is an equivalent of the only stratified sampling
with(sa_df2, my_table(.cluster, .strata))

#>    1  2  3 4  5  6  7  8  9 10    
#> 0 10 10 11 9 10 10 10 10 10 10 100
#>   10 10 11 9 10 10 10 10 10 10 100

# Structure of the three repetedly sampled datasets
get_bootstrap_permutation(sa_df2, 3) %>%
  map(~ tibble(ids = .x) %>% left_join(sa_df2, "ids")) %>%
  map(~ with(.x, my_table(.cluster, .strata)))

#> [[1]]
#>    1  2  3 4  5  6  7  8  9 10    
#> 0 10 10 11 9 10 10 10 10 10 10 100
#>   10 10 11 9 10 10 10 10 10 10 100
#> 
#> [[2]]
#>    1  2  3 4  5  6  7  8  9 10    
#> 0 10 10 11 9 10 10 10 10 10 10 100
#>   10 10 11 9 10 10 10 10 10 10 100
#> 
#> [[3]]
#>    1  2  3 4  5  6  7  8  9 10    
#> 0 10 10 11 9 10 10 10 10 10 10 100
#>   10 10 11 9 10 10 10 10 10 10 100
#> 

# Example of sampling with clusters and stratas,
# where data contains only one strata equall to 0
# This is an equivalent of the only clustered sampling
with(sa_df3, my_table(.cluster, .strata))

#>     1    
#> 1  15  15
#> 2  30  30
#> 3   5   5
#> 4  40  40
#> 5  10  10
#>   100 100

# Structure of the three repetedly sampled datasets
get_bootstrap_permutation(sa_df3, 3) %>%
  map(~ tibble(ids = .x) %>% left_join(sa_df3, "ids")) %>%
  map(~ with(.x, my_table(.cluster, .strata)))

#> [[1]]
#>    1   
#> 1 30 30
#> 2 30 30
#> 3  5  5
#> 5 10 10
#>   75 75
#> 
#> [[2]]
#>    1   
#> 1 15 15
#> 3 15 15
#> 4 40 40
#>   70 70
#> 
#> [[3]]
#>     1    
#> 1  15  15
#> 3  10  10
#> 4  80  80
#>   105 105
#> 

# Comparing old bootstrapping logic with the new one
df <- tibble::tibble(id = 1:30, group = sort(rep(1:3, 10))) %>%
  mutate(ids = id, .strata = group, .cluster = 1)

set.seed(1)
old <- get_bootstrap_permutation_old(df, 3) %>%
  map(~ sort(.x))
set.seed(1)
new <- get_bootstrap_permutation(df, 3) %>%
  map(~ sort(.x))
all_equal(old, new)

#> [1] TRUE