Get multiple imputation mean statistics

Get multiple imputation mean statistics

get_mi_mean(.data, stat_var, weight_var = NULL, ids = ".id",
  imps = ".imp", summary = TRUE, use_random = FALSE)

Arguments

.data
dataframe that could be converted to the mids object to perform calculations on the multiple imputations data frame.

stat_var
variable to apply statistics

weight_var
variable for weighting. NULL by default.

ids, imps
names of the columns wit the observation-specific ID and number of imputation.

summary
return a summary of the pooled analysis.

use_random
harmonise imputation of the unequal size randomly. By default if FALSE, therefore, to equalise sizes of the imputed samples, simplu n-first observations is used.

Examples

library(dplyr)
dep <- "price"            # Dependent variable
indep <- c("mpg", "headroom", "trunk",
           "weight", "length", "turn",
           "displacement", "gear_ratio",
           "foreign")     # independent variables
weight <- "weight"        # weight variable
extrar <- "displacement"  # any extra variable to bring from period 0 data
bt_groups <- c("psu")     # Grouping variable for bootsrapping.
n_nearest <- 5    # numebr of the nearest observations to drow a random match
set.seed(11344)
imputation <-
  lassopmm(
    source = p_0_exmple,
    target = p_1_exmple,
    dep_var = dep,
    indep_var = indep,
    weight_var = weight,
    group_boot_var = bt_groups,
    extra_var = extrar,
    n_boot = 5, # Numebr of bootstrap iterations
    n_near = 1)

# Getting results of the averages
get_mi_mean(imputation, stat_var = "price_source", use_random = F)

#> # A tibble: 1 x 15
#>   variable estimate std.error statistic    df p.value `2.5 %` `97.5 %`    riv
#>   <chr>       <dbl>     <dbl>     <dbl> <dbl>   <dbl>   <dbl>    <dbl>  <dbl>
#> 1 price_s~    6127.     1344.      4.56  4.35 0.00848   2512.    9743. 0.0676
#> # ... with 6 more variables: lambda <dbl>, fmi <dbl>, ubar <dbl>, b <dbl>,
#> #   t <dbl>, dfcom <int>

# Same by with weights in means
get_mi_mean(imputation, stat_var = "price_source", weight_var = "weight")

#> # A tibble: 1 x 15
#>   variable estimate std.error statistic    df p.value `2.5 %` `97.5 %`   riv
#>   <chr>       <dbl>     <dbl>     <dbl> <dbl>   <dbl>   <dbl>    <dbl> <dbl>
#> 1 price_s~    6927.     1615.      4.29  4.13  0.0119   2499.   11356. 0.117
#> # ... with 6 more variables: lambda <dbl>, fmi <dbl>, ubar <dbl>, b <dbl>,
#> #   t <dbl>, dfcom <int>